[hotfix] 홈페이지 오류 수정 (#26)
* style: navbar 스타일 및 아이콘 사이즈 수정(LUP-46)

* feat: modal component 구현(LUP-46)

* fix: login 컴포넌트 수정

* fix: PetInfo 컴포넌트 이름 수정

* fix: home 페이지 오류 수정

* hotfix: 브랜치 충돌 수정
