+++ frontend/src/components/common/Layout/index.tsx
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled';
-
 import Navbar from '../Navbar';
 import { Flex } from '../Flex';
 import { Header } from '../Header';
@@ -7,8 +6,8 @@ import { Header } from '../Header';
 export const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
     <Wrapper direction="column" align="center">
-      <Header />
-      <MobileWrapper direction="column">
+      <MobileWrapper width={375} direction="column">
+        <Header />
         <Flex padding="0 10px">{children}</Flex>
         <Navbar />
       </MobileWrapper>
