+++ frontend/src/components/common/Modal/index.tsx
@@ -1,3 +1,2 @@
 export { default as Modal } from './Modal';
-
 export * from './Modal';
